# HotelBookingDetails/HotelBookingDetails.WasmClient/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HotelBookingDetails.WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ReservedRoom": {
      "get": {
        "tags": [
          "ReservedRoom"
        ],
        "summary": "Запрос возвращающий полную информацию обо всех бронированиях",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservedRooms"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservedRooms"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservedRooms"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ReservedRoom"
        ],
        "summary": "Запрос на добавления бронирования",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservedRoomsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservedRoomsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReservedRoomsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ReservedRoom/{id}": {
      "get": {
        "tags": [
          "ReservedRoom"
        ],
        "summary": "Запрос информации о обронировании по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReservedRooms"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservedRooms"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservedRooms"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ReservedRoom"
        ],
        "summary": "Запрос на изменение бронирования по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservedRoomsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservedRoomsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReservedRoomsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ReservedRoom"
        ],
        "summary": "Удаление бронирования по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ReservedRoom/get_all_client_in_hotel/{hotelId}": {
      "get": {
        "tags": [
          "ReservedRoom"
        ],
        "summary": "Возвращает всех клиентов отеля",
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ReservedRoom/get_all_free_rooms/{city}": {
      "get": {
        "tags": [
          "ReservedRoom"
        ],
        "summary": "Возвращает все свободные комнаты в выбранном городе",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Room": {
      "get": {
        "tags": [
          "Room"
        ],
        "summary": "Запрос возвращающий список всех комнат",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Room"
        ],
        "summary": "Запрос на добавления комнаты",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Room/{id}": {
      "get": {
        "tags": [
          "Room"
        ],
        "summary": "Запрос комнаты по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Room"
        ],
        "summary": "Запрос на изменение комнаты по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Room"
        ],
        "summary": "Удаление комнаты по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Client": {
        "required": [
          "birthday",
          "fullName",
          "id",
          "passportData"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "minLength": 1,
            "type": "string"
          },
          "passportData": {
            "$ref": "#/components/schemas/Passport"
          },
          "birthday": {
            "$ref": "#/components/schemas/DateOnly"
          }
        },
        "additionalProperties": false
      },
...
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Hotel": {
        "required": [
          "address",
          "city",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "address": {
            "minLength": 1,
            "type": "string"
          },
          "city": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Passport": {
        "required": [
          "id",
          "number",
          "series"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "series": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ReservedRooms": {
        "required": [
          "client",
          "dateArrival",
          "id",
          "period",
          "room"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          },
          "dateArrival": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "dateDeparture": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "period": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ReservedRoomsDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "roomId": {
            "type": "integer",
            "format": "int32"
          },
          "dateArrival": {
            "type": "string",
            "nullable": true
          },
          "dateDeparture": {
            "type": "string",
            "nullable": true
          },
          "period": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Room": {
        "required": [
          "capacity",
          "cost",
          "hotel",
          "id",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/TypeRoom"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "integer",
            "format": "int32"
          },
          "hotel": {
            "$ref": "#/components/schemas/Hotel"
          }
        },
        "additionalProperties": false
      },
      "RoomDto": {
        "type": "object",
        "properties": {
          "typeId": {
            "type": "integer",
            "format": "int32"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "integer",
            "format": "int32"
          },
          "hotelId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TypeRoom": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}